+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 ## [Unreleased][Unreleased]
 ### Added
 - [Feature] Add a `shouldHideOnClick` prop to `OverlayWrapper` to prevent it from always closing when clicking outside the overlay. (#654)
+- [Feature] Add `onShow` and `onHide` props to `OverlayWrapper`. (#653)
 - [Feature] Add a `centerGroup` prop to `ButtonRow`. (#651)
 - [Feature] Expose a `hideOverlay` function in `OverlayWrapper` as context that allows children to hide the overlay. (#657)